Fix returning and displaying trips without stopEvents
Without the `v-if`, the site didn't render because of accessing `arrivalDelays` on a `null` object. Also, to get ALL trips (not just the ones that we caught data for), we need to `GROUP BY` a `trips` field (not a `stopEvents` field) and do a join that always fully includes that table. (Also unrelatedly fixing a path and a tab.)